feat:  채팅방 삭제시 , 메시지 먼저 다 지워지고 삭제 되도록

+++ backend/src/main/java/com/example/backend/controller/chat/ChatController.java
@@ -91,11 +91,23 @@ public class ChatController {
     @DeleteMapping("delete/{roomId}")
     public ResponseEntity<Map<String, Object>> deleteChatRoom(@PathVariable String roomId) {
 
-        boolean run = chatService.deleteChatRoom(roomId);
-        if (run) {
-            return ResponseEntity.ok().body(Map.of("message", Map.of("type", "success", "content", "채팅방 삭제 완료되었습니다.")));
-        } else {
-            return ResponseEntity.badRequest().body(Map.of("message", Map.of("type", "success", "content", "존재하지 않는 채팅방 입니다.")));
-        }
+
+        //채팅 전에 메세지를 먼저 삭제하고 > 채팅방을 삭제
+        // 채팅방 삭제전 , 미리  채팅 메시지삭제 하는 코드
+        boolean messageRemoved = chatService.deleteMessageAll(roomId);
+        boolean chatRemoved;
+        // 실행 여분데
+        if (messageRemoved) {
+            chatRemoved = chatService.deleteChatRoom(roomId);
+            if (chatRemoved) {
+                return ResponseEntity.ok().body(Map.of("message", Map.of("type", "success", "content", "채팅방 삭제 완료되었습니다.")));
+            } else {
+                return ResponseEntity.badRequest().body(Map.of("message", Map.of("type", "warning", "content", "존재하지 않는 채팅방 입니다.")));
+            }
+        } else
+            return ResponseEntity.badRequest().body(Map.of("message", Map.of("type", "success", "content", "뭘 삭제하려는 겁니까")));
+
+
     }
+
 }
